split run page and playground component
